@@ -1,13 +1,15 @@
-import * as React from "react";
+import { FunctionComponent, useState } from "react";
 
-//
-class Task1 extends React.Component {
-  state = {
-    data: ["Item 1", "Item 2", "Item 3", "Item 4", "Item 5"]
-  };
-  render() {
-    return <div>#List goes here#</div>;
-  }
-}
+const Task1: FunctionComponent = () => {
+  const [items, setItems] = useState([
+    "Item 1",
+    "Item 2",
+    "Item 3",
+    "Item 4",
+    "Item 5",
+  ]);
+
+  return <div>#List goes here#</div>;
+};
 
 export default Task1;
