@@ -1,22 +1,18 @@
-import * as React from "react";
+import { FunctionComponent, useState } from "react";
 
-//
-class Task3 extends React.Component {
-  state = {
-    inputValue: "",
-    data: [""]
-  };
-  render() {
-    return (
-      <div>
-        <label>Search Input: </label>
-        <input />
-        <br />
-        <br />
-        #List goes here#
-      </div>
-    );
-  }
-}
+const Task3: FunctionComponent = () => {
+  const [inputValue, setInputValue] = useState("");
+  const [items, setItems] = useState([]);
+
+  return (
+    <div>
+      <label>Search Input: </label>
+      <input />
+      <br />
+      <br />
+      #List goes here#
+    </div>
+  );
+};
 
 export default Task3;
