@@ -1,18 +1,18 @@
-import * as React from "react";
+import { FunctionComponent } from "react";
 
 /*
-     * The IProps interface defines the props in order to typescript to typecheck for errors.
-     * 
-     * If you would like to proceed without defining types do the following: 
-     * const Input: React.SFC<any> = (props) => {
-     *                        ^^^
-     * 
-     * and remove the IProps interface
-*/
+ * The ItemProps interface defines the types for the components props.
+ *
+ * If you would like to proceed without defining types do the following:
+ * const Input: FunctionComponent<any> = (props) => {
+ *                                ^^^
+ *
+ * and remove the ItemProps interface
+ */
 
-//
-interface IProps {}
-const Item: React.SFC<IProps> = props => {
+interface ItemProps {}
+
+const Item: FunctionComponent<ItemProps> = (props) => {
   return <li>#Item goes here#</li>;
 };
 
