@@ -1,21 +1,21 @@
-import * as React from "react";
+import { FunctionComponent } from "react";
 
 // Components
 import Item from "./Item";
 
 /*
-     * The IProps interface defines the props in order to typescript to typecheck for errors.
-     * 
-     * If you would like to proceed without defining types do the following: 
-     * const Input: React.SFC<any> = (props) => {
-     *                        ^^^
-     * 
-     * and remove the IProps interface
-*/
+ * The ListProps interface defines the types for the components props.
+ *
+ * If you would like to proceed without defining types do the following:
+ * const Input: FunctionComponent<any> = (props) => {
+ *                                ^^^
+ *
+ * and remove the ListProps interface
+ */
 
-//
-interface IProps {}
-const List: React.SFC<IProps> = props => {
+interface ListProps {}
+
+const List: FunctionComponent<ListProps> = (props) => {
   return <div>#List goes here#</div>;
 };
 
