@@ -1,20 +1,17 @@
-import * as React from "react";
+import { FunctionComponent } from "react";
 
 // Components
 import Input from "./components/Input";
 import List from "./components/List";
 
-//
-class Task4 extends React.Component {
-  render() {
-    return (
-      <div>
-        <Input />
-        <br />
-        <List />
-      </div>
-    );
-  }
-}
+const Task4: FunctionComponent = () => {
+  return (
+    <div>
+      <Input />
+      <br />
+      <List />
+    </div>
+  );
+};
 
 export default Task4;
