@@ -1,26 +1,18 @@
 import * as React from "react";
 
+// Style
+import "./index.scss";
+
 // Components
 /*
 * Create the components you need in the components folder.
 * You may find inspiration in task 4
 */
-
-// Style
-/*
-* Use this less file to create your own styles.
-*/
-import "./index.less";
-
-//
-class Task6 extends React.Component {
-  render() {
-    return (
-      <div>
-        #Code goes here#
-      </div>
-    );
-  }
+ 
+const Task6: React.FunctionComponent = () => {
+  return (        <div>
+    #Code goes here#
+  </div>);
 }
-
+ 
 export default Task6;
